# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Memora API (mock)
  version: "1.0.0"
  description: |
    API mock do Memora — Express 5, estado em memória (reinicia com o processo, sem banco de
    dados). Reproduz as regras de negócio do Escopo Consolidado v2 para servir de backend ao
    protótipo navegável; não é a API de produção.

    Todas as rotas de negócio vivem sob `/v1`. A autenticação é por sessão simples (token opaco
    em `Authorization: Bearer <token>`), sem JWT nem refresh — ver
    [ADR-0008](./adr/0008-autenticacao-por-sessao-simples.md).

    Cada área do produto (aluno, responsável, B2B, back-office, parceiro) faz login com sua
    própria rota `POST .../auth/login` e recebe um token cujo `role` restringe o que pode
    acessar (`requireRole` em `backend/src/lib/auth.js`).
  contact:
    name: Memora — engenharia
  license:
    name: Uso interno
servers:
  - url: http://localhost:3180/v1
    description: Desenvolvimento local (porta configurável via `PORT`)
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: >
        Token de sessão retornado por qualquer rota `auth/login`. Sessões expiram em 30 dias
        sem uso (deslizante); usar em `me`, `parental`, `b2b`, `admin` e `partner` é mutuamente
        exclusivo por `role`.
  parameters:
    deckId:
      name: id
      in: path
      required: true
      schema: { type: string }
      description: ID do deck.
    cardId:
      name: cardId
      in: path
      required: true
      schema: { type: string }
  responses:
...
    NotFound:
      description: Recurso não encontrado (ou não pertence ao usuário autenticado).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Validation:
      description: Corpo da requisição inválido.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, description: "Slug estável da causa, ex.: validation, not_found, guardian_required." }
        message: { type: string, description: "Mensagem em PT-BR, pronta para exibição." }
      additionalProperties: true
      description: Formato uniforme de erro em toda a API (ver `fail()` em `backend/src/lib/helpers.js`).
...
    Deck:
      type: object
      properties:
        id: { type: string }
        ownerId: { type: string }
        name: { type: string }
        categoryId: { type: string }
        categoryName: { type: string }
        tags: { type: array, items: { type: string } }
        difficulty: { type: string, enum: [facil, medio, dificil] }
        lang: { type: string }
        difficult: { type: boolean }
        focus: { type: boolean }
        complementary: { type: boolean }
        pinned: { type: boolean }
        folderId: { type: string, nullable: true }
        source: { type: object, additionalProperties: true }
        createdAt: { type: string, format: date-time }
        lastStudiedAt: { type: string, format: date-time, nullable: true }
        publication: { type: object, additionalProperties: true }
        cardCount: { type: integer }
        progress:
          type: object
          properties:
            total: { type: integer }
            due: { type: integer }
            mastered: { type: integer }
            learning: { type: integer }
            offlineAvailable: { type: boolean }
            offlineReason: { type: string, nullable: true }
    Folder:
      type: object
      properties:
        id: { type: string }
        ownerId: { type: string }
        name: { type: string }
        parentId: { type: string, nullable: true }
        depth: { type: integer }
        shared: { type: boolean }
        members:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              name: { type: string }
              contact: { type: string }
              role: { type: string }
              since: { type: string, format: date-time }
        deckIds: { type: array, items: { type: string } }
...
    Publication:
      type: object
      description: Submissão de um deck para publicação na comunidade (fila de moderação).
      properties:
        id: { type: string }
        deckId: { type: string }
        authorId: { type: string }
        authorName: { type: string }
        deckName: { type: string }
        version: { type: integer }
        status: { type: string, enum: [em_triagem, em_revisao, aprovado, rejeitado] }
        plan: { type: string }
        submittedAt: { type: string, format: date-time }
        decidedAt: { type: string, format: date-time, nullable: true }
        slaHours: { type: integer }
        risk:
          type: object
          properties:
            score: { type: integer }
            level: { type: string, enum: [baixo, medio, alto] }
            factors: { type: array, items: { type: string } }
        reviewType: { type: string, enum: [integral, amostral] }
        decision:
          type: object
          nullable: true
          properties:
            by: { type: string }
            byName: { type: string }
            decision: { type: string, enum: [aprovado, rejeitado] }
            reasonCategory: { type: string, nullable: true }
            excerpt: { type: string, nullable: true }
            appealDeadline: { type: string, format: date-time, nullable: true }
        sla:
          type: object
          properties:
            pct: { type: integer }
            dueAt: { type: string, format: date-time, nullable: true }
            alert: { type: boolean }
            hoursLeft: { type: integer }
    Appeal:
      type: object
      properties:
        id: { type: string }
        publicationId: { type: string }
        authorName: { type: string }
        filedAt: { type: string, format: date-time }
        deadline: { type: string, format: date-time }
        text: { type: string }
        status: { type: string, enum: [em_analise, aceita, mantida] }
        reviewer: { type: string, nullable: true }
        originalReviewer: { type: string, nullable: true }
...
paths:
...
  /decks:
    get:
      tags: [Decks]
      summary: Meus decks, favoritos, pastas e limites do plano
      responses: { "200": { description: OK } }
    post:
      tags: [Decks]
      summary: Cria um deck vazio
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties: { name: { type: string }, categoryId: { type: string }, tags: { type: array, items: { type: string } }, difficulty: { type: string }, folderId: { type: string, nullable: true } }
      responses:
        "201": { description: Criado, content: { application/json: { schema: { type: object, properties: { deck: { $ref: '#/components/schemas/Deck' } } } } } }
        "422": { $ref: '#/components/responses/Validation' }
  /decks/{id}:
    get:
      tags: [Decks]
      summary: Detalhe do deck, cards e menu de ações disponíveis
      parameters: [{ $ref: '#/components/parameters/deckId' }]
      responses:
        "200": { description: OK }
        "404": { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Decks]
      summary: Atualiza flags (difícil/foco/complementar/fixado), nome, categoria, tags ou pasta
      parameters: [{ $ref: '#/components/parameters/deckId' }]
      responses:
        "200": { description: OK }
        "404": { $ref: '#/components/responses/NotFound' }
        "409": { description: "`pin_limit` — limite de decks fixados do plano atingido." }
    delete:
      tags: [Decks]
      summary: Exclui o deck (soft delete)
      parameters: [{ $ref: '#/components/parameters/deckId' }]
      responses:
        "200": { description: OK }
        "404": { $ref: '#/components/responses/NotFound' }
  /decks/{id}/cards:
    post:
      tags: [Decks]
      summary: Adiciona um card manualmente (até 2.000 por deck)
      parameters: [{ $ref: '#/components/parameters/deckId' }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [front, back]
              properties: { front: { type: string }, back: { type: string }, type: { type: string }, media: { type: object, nullable: true } }
      responses:
        "201": { description: Criado }
        "409": { description: deck_limit }
        "422": { $ref: '#/components/responses/Validation' }
  /decks/{id}/cards/{cardId}:
    patch:
      tags: [Decks]
      summary: Edita frente/verso do card
      parameters: [{ $ref: '#/components/parameters/deckId' }, { $ref: '#/components/parameters/cardId' }]
      responses:
        "200": { description: OK }
        "404": { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Decks]
      summary: Remove o card (soft delete)
      parameters: [{ $ref: '#/components/parameters/deckId' }, { $ref: '#/components/parameters/cardId' }]
      responses:
        "200": { description: OK }
        "404": { $ref: '#/components/responses/NotFound' }
  /decks/{id}/publish:
    post:
      tags: [Decks]
      summary: Submete o deck para publicação na comunidade
      description: >
        Exige e-mail/telefone confirmados, liberação do responsável se menor (`guardian_required`),
        e aceite da política de conteúdo na primeira vez (`policy_acceptance_required`, HTTP 428).
      parameters: [{ $ref: '#/components/parameters/deckId' }]
      requestBody:
        content: { application/json: { schema: { type: object, properties: { acceptPolicy: { type: boolean } } } } }
      responses:
        "201": { description: Enviado para triagem, content: { application/json: { schema: { type: object, properties: { publication: { $ref: '#/components/schemas/Publication' } } } } } }
        "403": { description: "guardian_required ou verification_required ou account_deactivated" }
        "422": { description: empty_deck }
        "428": { description: policy_acceptance_required }
  /publications/{id}/appeal:
    post:
      tags: [Decks]
      summary: Contesta uma publicação rejeitada (janela de 30 dias, revisor diferente do original)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        content: { application/json: { schema: { type: object, required: [text], properties: { text: { type: string } } } } }
      responses:
        "201": { description: Criada, content: { application/json: { schema: { type: object, properties: { appeal: { $ref: '#/components/schemas/Appeal' } } } } } }
        "404": { $ref: '#/components/responses/NotFound' }
        "409": { description: "not_rejected · appeal_window_closed · appeal_exists" }
  /decks/{id}/export:
    post:
      tags: [Decks]
      summary: Exporta o deck (PDF é entregue como texto no protótipo)
      parameters: [{ $ref: '#/components/parameters/deckId' }]
      requestBody:
        content: { application/json: { schema: { type: object, properties: { format: { type: string, enum: [pdf, txt] } } } } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: { fileName: { type: string }, format: { type: string }, url: { type: string, description: "data: URL com o conteúdo." } }
  /folders:
    post:
      tags: [Decks]
      summary: Cria uma pasta (até 3 níveis de profundidade)
      requestBody:
        content: { application/json: { schema: { type: object, required: [name], properties: { name: { type: string }, parentId: { type: string, nullable: true } } } } }
      responses:
        "201": { description: Criada, content: { application/json: { schema: { type: object, properties: { folder: { $ref: '#/components/schemas/Folder' } } } } } }
        "409": { description: folder_depth }
  /folders/{id}:
    patch:
      tags: [Decks]
      summary: Renomeia a pasta
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: OK }
        "404": { $ref: '#/components/responses/NotFound' }
  /folders/{id}/members:
    post:
      tags: [Decks]
      summary: Adiciona um membro externo à pasta (até 50; exige liberação do responsável se menor)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        content: { application/json: { schema: { type: object, required: [contact], properties: { name: { type: string }, contact: { type: string } } } } }
      responses:
        "201": { description: OK }
        "403": { description: guardian_required }
        "409": { description: members_limit }
  /folders/{id}/members/{memberId}:
    delete:
      tags: [Decks]
      summary: Remove um membro da pasta
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: memberId, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: OK }
        "404": { $ref: '#/components/responses/NotFound' }

  # --------------------------------------------------------------- Geração
...